limit of the DNA template amount for outerPCR
Before, I have used all 1st PCR product (no matter what is the
concentration) as the outerPCR template, which seems to work fine. But
I have thought about it, that would result in the primer depletion in a
few cycles of the outerPCR and therefore introducing a small fraction
of unindexed 1stPCR amplicon product to the final sequencing library.
The side effect is the loss of sequencing reads on this type of
library. (1pg-10ng of low complexity DNA template in 50uL PCR reaction
was suggested by NEB, ref:
https://www.neb.com/tools-and-resources/troubleshooting-guides/pcr-troub
leshooting-guide )